Binary runs server with `sundial listen`
