+++ package.json
@@ -19,6 +19,7 @@
     "axios": "^0.27.2",
     "commander": "^11.1.0",
     "crypto": "^1.0.1",
+    "express": "^4.18.2",
     "nanoid": "^5.0.2",
     "node-cron": "^3.0.2"
   },
