+++ src/commands/index.js
@@ -2,10 +2,12 @@ import { run } from './run.js';
 import { discover } from './discover.js';
 import { update } from './update.js';
 import { register } from './register.js';
+import { listen } from './listen.js';
 
 export default {
   run,
   discover,
   update,
-  register
+  register,
+  listen,
 };
